tiny refactoring and fixes in TErrorTracker
ref:f08bdc6123481ba48a5f5e1443623a0415ccacc4

+++ catboost/libs/algo/train.h
@@ -9,4 +9,4 @@ using TTrainOneIterationFunc = std::function<void(const TDataset& learnData,
 
 TTrainOneIterationFunc GetOneIterationFunc(ELossFunction lossFunction);
 
-TErrorTracker BuildErrorTracker(EMetricBestValue bestValueType, float bestPossibleValue, bool hasTest, TLearnContext* ctx);
+TErrorTracker BuildErrorTracker(EMetricBestValue bestValueType, double bestPossibleValue, bool hasTest, TLearnContext* ctx);
